Move all tests together in their modules
