+++ lib/path/__tests__/parsers/directory.test.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const test = require('ava')
-const parseDirectory = require('../directory')
+const parseDirectory = require('../../parsers/directory')
 
 test('should not update token if already analyzed', async t => {
   const token = {
