+++ lib/path/__tests__/parsers/file.test.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const test = require('ava')
-const parseFile = require('../file')
+const parseFile = require('../../parsers/file')
 
 const options = {
   digestAlgorithm: 'sha384'
