+++ lib/path/__tests__/parsers/type.test.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const parseType = require('../type')
+const parseType = require('../../parsers/type')
 
 test('should not update token if already analyzed', t => {
   const token = {
