+++ lib/url/__tests__/parsers/archive.test.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const parseAtom = require('../atom')
+const parseArchive = require('../../parsers/archive')
 
 test('should not update token if already analyzed', async t => {
   const token = {
@@ -7,7 +7,7 @@ test('should not update token if already analyzed', async t => {
   }
   const save = Object.assign({}, token)
 
-  const ret = await parseAtom(token)
+  const ret = await parseArchive(token)
 
   t.is(ret, false)
   t.deepEqual(token, save)
