+++ lib/url/__tests__/parsers/atom.test.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const parseArchive = require('../archive')
+const parseAtom = require('../../parsers/atom')
 
 test('should not update token if already analyzed', async t => {
   const token = {
@@ -7,7 +7,7 @@ test('should not update token if already analyzed', async t => {
   }
   const save = Object.assign({}, token)
 
-  const ret = await parseArchive(token)
+  const ret = await parseAtom(token)
 
   t.is(ret, false)
   t.deepEqual(token, save)
