+++ lib/url/__tests__/parsers/file.test.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const parseHtml = require('../html')
+const parseFile = require('../../parsers/file')
 
 test('should not update token if already analyzed', async t => {
   const token = {
@@ -7,7 +7,7 @@ test('should not update token if already analyzed', async t => {
   }
   const save = Object.assign({}, token)
 
-  const ret = await parseHtml(token)
+  const ret = await parseFile(token)
 
   t.is(ret, false)
   t.deepEqual(token, save)
