+++ lib/url/__tests__/parsers/html.test.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const parseFile = require('../file')
+const parseHtml = require('../../parsers/html')
 
 test('should not update token if already analyzed', async t => {
   const token = {
@@ -7,7 +7,7 @@ test('should not update token if already analyzed', async t => {
   }
   const save = Object.assign({}, token)
 
-  const ret = await parseFile(token)
+  const ret = await parseHtml(token)
 
   t.is(ret, false)
   t.deepEqual(token, save)
