+++ lib/url/__tests__/parsers/name.test.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const parseName = require('../name')
+const parseName = require('../../parsers/name')
 
 test('should not update token if already analyzed', t => {
   const token = {
