+++ lib/url/__tests__/parsers/types.test.js
@@ -1,5 +1,5 @@
 const test = require('ava')
-const parseTypes = require('../types')
+const parseTypes = require('../../parsers/types')
 
 test('should not update token if already analyzed', async t => {
   const token = {
